%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  getCacheRootPath.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCacheRootPath}
\alias{getCacheRootPath.default}
\alias{getCacheRootPath}

\title{Gets the root path to the file cache directory}

\description{
 Gets the root path to the file cache directory.
}

\usage{\method{getCacheRootPath}{default}(defaultPath="~/.Rcache", ...)}

\arguments{
  \item{defaultPath}{The default path.}
  \item{...}{Not used.}
}

\value{
  Returns the path as a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
 \code{\link[R.cache:setCacheRootPath.]{*setCacheRootPath}()}.
}




\keyword{programming}
\keyword{IO}
\keyword{internal}
