%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  setCacheRootPath.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setCacheRootPath}
\alias{setCacheRootPath.default}
\alias{setCacheRootPath}

\title{Sets the root path to the file cache directory}

\description{
 Sets the root path to the file cache directory.
}

\usage{
\method{setCacheRootPath}{default}(path=NULL, ...)
}

\arguments{
  \item{path}{The path.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) the old root path.
}

\author{Henrik Bengtsson}

\seealso{
 \code{\link{getCacheRootPath}}().
}



\keyword{programming}
\keyword{IO}
